refactor namin
